+++ openmixup/models/heads/__init__.py
@@ -1,5 +1,6 @@
 from .cls_head import ClsHead
 from .cls_mixup_head import ClsMixupHead
+from .cls_mlp_head import EfficientFormerClsHead, MetaFormerClsHead
 from .contrastive_head import ContrastiveHead, HCRHead
 from .latent_pred_head import LatentPredictHead, LatentClsHead, LatentCrossCorrelationHead, MoCoV3Head
 from .mim_head import A2MIMHead, MAEPretrainHead, MAEFinetuneHead, MAELinprobeHead, SimMIMHead
@@ -13,6 +14,7 @@ from .vision_transformer_head import VisionTransformerClsHead
 
 __all__ = [
     'A2MIMHead', 'BEiTHead', 'CAEHead', 'ClsHead', 'ClsMixupHead', 'ContrastiveHead', 'HCRHead',
+    'EfficientFormerClsHead', 'MetaFormerClsHead',
     'LatentPredictHead', 'LatentClsHead', 'LatentCrossCorrelationHead',
     'MoCoV3Head', 'MAEPretrainHead', 'MAELinprobeHead', 'MAEFinetuneHead', 'MAELinprobeHead',
     'MultiClsHead', 'MultiPrototypes', 'NormLinearClsHead', 'PixelMixBlock', 'RegHead',