+++ openmixup/models/heads/cls_head.py
@@ -10,15 +10,15 @@ from ..registry import HEADS
 from ..builder import build_loss
 
 
-@HEADS.register_module
-class ClsHead(BaseModule):
-    """Simplest classifier head, with only one fc layer.
+class BaseClsHead(BaseModule):
+    """The Basic Classification Head.
        *** Mixup and multi-label classification are supported ***
-    
+
     Args:
         with_avg_pool (bool): Whether to use GAP before this head.
         loss (dict): Config of classification loss.
         in_channels (int): Number of channels in the input feature map.
+        hidden_dim (int or None): Number of hidden channels in a MLP ClsHead.
         num_classes (int): Number of categories excluding the category.
         multi_label (bool): Whether to use one_hot like labels (requiring the
             multi-label classification loss). Notice that we support the
@@ -26,25 +26,30 @@ class ClsHead(BaseModule):
         finetune (bool): Whether to use the finetune mode of ViTs.
         aug_test (bool): Whether to perform test time augmentations.
         frozen (bool): Whether to freeze the parameters.
+        **kwargs: Support other args for ClsHead variants.
     """
 
     def __init__(self,
                  with_avg_pool=False,
                  loss=dict(type='CrossEntropyLoss', loss_weight=1.0),
                  in_channels=2048,
+                 hidden_dim=None,
                  num_classes=1000,
                  multi_label=False,
                  finetune=False,
                  aug_test=False,
                  frozen=False,
-                 init_cfg=None):
-        super(ClsHead, self).__init__(init_cfg=init_cfg)
+                 init_cfg=None,
+                 **kwargs):
+        super(BaseClsHead, self).__init__(init_cfg=init_cfg)
         self.with_avg_pool = with_avg_pool
         self.in_channels = in_channels
+        self.hidden_dim = hidden_dim
         self.num_classes = num_classes
         self.multi_label = multi_label
         self.finetune = finetune
         self.aug_test = aug_test
+        self.frozen = frozen
 
         # loss
         if loss is not None:
@@ -55,12 +60,9 @@ class ClsHead(BaseModule):
             loss = dict(type='CrossEntropyLoss', loss_weight=1.0)
             self.criterion = build_loss(loss)
         self.processor = getattr(self.criterion, "processor", None)
-        # fc layer
-        self.fc = None
-        if num_classes is not None:
-            self.fc = nn.Linear(in_channels, num_classes)
-        if frozen:
-            self.frozen()
+
+        # TODO: build a classification head
+
         # post-process for inference
         post_process = getattr(self.criterion, "post_process", "none")
         if post_process == "softmax":
@@ -68,45 +70,16 @@ class ClsHead(BaseModule):
         else:
             self.post_process = nn.Identity()
 
-    def frozen(self):
-        if self.fc is None:
-            return
-        self.fc.eval()
-        for param in self.fc.parameters():
-            param.requires_grad = False
+    def _freeze(self):
+        """ freeze classification heads """
+        raise NotImplementedError
 
     def init_weights(self, init_linear='normal', std=0.01, bias=0.):
-        if self.fc is None or self.init_cfg is not None:
-            super(ClsHead, self).init_weights()
-            return
-        assert init_linear in ['normal', 'kaiming', 'trunc_normal'], \
-            "Undefined init_linear: {}".format(init_linear)
-        if self.finetune:  # finetune for ViTs
-            std = 2e-5
-            init_linear = 'trunc_normal'
-        for m in self.modules():
-            if isinstance(m, nn.Linear):
-                if init_linear == 'normal':
-                    normal_init(m, std=std, bias=bias)
-                elif init_linear == 'kaiming':
-                    kaiming_init(m, mode='fan_in', nonlinearity='relu')
-                elif init_linear == 'trunc_normal':
-                    trunc_normal_init(m, std=std, bias=bias)
+        raise NotImplementedError
 
     def forward_head(self, x, post_process=False):
         """" forward cls head with x in a shape of (X, \*) """
-        if self.with_avg_pool:
-            if x.dim() == 3:
-                x = F.adaptive_avg_pool1d(x, 1).view(x.size(0), -1)
-            elif x.dim() == 4:
-                x = F.adaptive_avg_pool2d(x, 1).view(x.size(0), -1)
-            else:
-                assert x.dim() in [2, 3, 4], \
-                    "Tensor must has 2, 3 or 4 dims, got: {}".format(x.dim())
-        x = self.fc(x)
-        if post_process:
-            x = self.post_process(x)
-        return x
+        raise NotImplementedError
 
     def forward(self, x, post_process=False, **kwargs):
         """Inference.
@@ -122,8 +95,6 @@ class ClsHead(BaseModule):
             Tensor | list: The inference results.
         """
         assert isinstance(x, (tuple, list)) and len(x) >= 1
-        if self.fc is None:
-            return x
         # test-time augmentation
         if len(x) > 1 and self.aug_test:
             aug_pred = [self.forward_head(_x, post_process) for _x in x]
@@ -134,8 +105,8 @@ class ClsHead(BaseModule):
             return [self.forward_head(x[0], post_process)]
 
     def loss(self, cls_score, labels, **kwargs):
-        """" cls loss forward
-        
+        """" classification loss forward
+
         Args:
             cls_score (list): Score should be [tensor].
             labels (tuple or tensor): Labels should be tensor [N, \*] by default.
@@ -208,3 +179,73 @@ class ClsHead(BaseModule):
             losses['acc'] = accuracy(cls_score, labels[0])
             losses['acc_mix'] = accuracy_mixup(cls_score, labels)
         return losses
+
+
+@HEADS.register_module
+class ClsHead(BaseClsHead):
+    """Simplest classifier head, with only one fc layer.
+       *** Mixup and multi-label classification are supported ***
+
+    Args:
+        with_avg_pool (bool): Whether to use GAP before this head.
+        loss (dict): Config of classification loss.
+        in_channels (int): Number of channels in the input feature map.
+        hidden_dim (int or None): Number of hidden channels in a MLP ClsHead.
+        num_classes (int): Number of categories excluding the category.
+        multi_label (bool): Whether to use one_hot like labels (requiring the
+            multi-label classification loss). Notice that we support the
+            single-label cls task to use the multi-label cls loss.
+        finetune (bool): Whether to use the finetune mode of ViTs.
+        aug_test (bool): Whether to perform test time augmentations.
+        frozen (bool): Whether to freeze the parameters.
+    """
+
+    def __init__(self, **kwargs):
+        super(ClsHead, self).__init__(**kwargs)
+
+        # build a classification head
+        assert self.hidden_dim is None
+        if self.num_classes is not None:
+            self.fc = nn.Linear(self.in_channels, self.num_classes)
+        if self.frozen:
+            self._freeze()
+
+    def _freeze(self):
+        if self.fc is None:
+            return
+        self.fc.eval()
+        for param in self.fc.parameters():
+            param.requires_grad = False
+
+    def init_weights(self, init_linear='normal', std=0.01, bias=0.):
+        if self.init_cfg is not None:
+            super(ClsHead, self).init_weights()
+            return
+        assert init_linear in ['normal', 'kaiming', 'trunc_normal'], \
+            "Undefined init_linear: {}".format(init_linear)
+        if self.finetune:  # finetune for ViTs
+            std = 2e-5
+            init_linear = 'trunc_normal'
+        for m in self.modules():
+            if isinstance(m, nn.Linear):
+                if init_linear == 'normal':
+                    normal_init(m, std=std, bias=bias)
+                elif init_linear == 'kaiming':
+                    kaiming_init(m, mode='fan_in', nonlinearity='relu')
+                elif init_linear == 'trunc_normal':
+                    trunc_normal_init(m, std=std, bias=bias)
+
+    def forward_head(self, x, post_process=False):
+        """" forward cls head with x in a shape of (X, \*) """
+        if self.with_avg_pool:
+            if x.dim() == 3:
+                x = F.adaptive_avg_pool1d(x, 1).view(x.size(0), -1)
+            elif x.dim() == 4:
+                x = F.adaptive_avg_pool2d(x, 1).view(x.size(0), -1)
+            else:
+                assert x.dim() in [2, 3, 4], \
+                    "Tensor must has 2, 3 or 4 dims, got: {}".format(x.dim())
+        x = self.fc(x)
+        if post_process:
+            x = self.post_process(x)
+        return x